[INTEGRATED-1122] block management refactored, also possible on contentTypePages
