+++ Controller/PageController.php
@@ -36,8 +36,8 @@ class PageController extends Controller
     {
         $channel = $this->getSelectedChannel();
 
-        $builder = $this->getDocumentManager()->createQueryBuilder('IntegratedPageBundle:Page\Page');
-        $builder->field('channel.$id')->equals($channel->getId());
+        $builder = $this->getDocumentManager()->createQueryBuilder(Page::class)
+            ->field('channel.$id')->equals($channel->getId());
 
         $pagination = $this->getPaginator()->paginate(
             $builder,
