+++ Document/Page/Grid/Column.php
@@ -11,15 +11,12 @@
 
 namespace Integrated\Bundle\PageBundle\Document\Page\Grid;
 
-use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
 use Doctrine\Common\Collections\ArrayCollection;
 
 /**
  * Column document
  *
  * @author Ger Jan van den Bosch <gerjan@e-active.nl>
- *
- * @ODM\EmbeddedDocument
  */
 class Column implements ItemsInterface
 {
@@ -31,7 +28,6 @@ class Column implements ItemsInterface
 
     /**
      * @var Item[]
-     * @ODM\EmbedMany(targetDocument="Integrated\Bundle\PageBundle\Document\Page\Grid\Item")
      */
     protected $items;
