+++ Resources/config/doctrine/Page.ContentTypePage.mongodb.xml
@@ -1,0 +1,13 @@
+<doctrine-mongo-mapping xmlns="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping.xsd">
+
+    <document name="Integrated\Bundle\PageBundle\Document\Page\ContentTypePage">
+
+        <field fieldName="controllerService" type="string" />
+        <field fieldName="controllerAction" type="string" />
+
+        <reference-one field="contentType" target-document="Integrated\Bundle\ContentBundle\Document\ContentType\ContentType" />
+    </document>
+
+</doctrine-mongo-mapping>
