+++ Resources/config/doctrine/Page.Grid.Column.mongodb.xml
@@ -1,0 +1,13 @@
+<doctrine-mongo-mapping xmlns="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping.xsd">
+
+    <embedded-document name="Integrated\Bundle\PageBundle\Document\Page\Grid\Column">
+
+        <field fieldName="size" type="int" />
+
+        <embed-many field="items" target-document="Integrated\Bundle\PageBundle\Document\Page\Grid\Item" />
+
+    </embedded-document>
+
+</doctrine-mongo-mapping>
