+++ Resources/config/doctrine/Page.Grid.Item.mongodb.xml
@@ -1,0 +1,15 @@
+<doctrine-mongo-mapping xmlns="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping.xsd">
+
+    <embedded-document name="Integrated\Bundle\PageBundle\Document\Page\Grid\Item">
+
+        <field fieldName="order" type="int" />
+
+        <embed-one field="row" target-document="Integrated\Bundle\PageBundle\Document\Page\Grid\Row" />
+
+        <reference-one field="block" target-document="Integrated\Bundle\BlockBundle\Document\Block\Block" />
+
+    </embedded-document>
+
+</doctrine-mongo-mapping>
