+++ Resources/config/doctrine/Page.Grid.Row.mongodb.xml
@@ -1,0 +1,11 @@
+<doctrine-mongo-mapping xmlns="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xsi:schemaLocation="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping.xsd">
+
+    <embedded-document name="Integrated\Bundle\PageBundle\Document\Page\Grid\Row">
+
+        <embed-many field="columns" target-document="Integrated\Bundle\PageBundle\Document\Page\Grid\Column" strategy="set" />
+
+    </embedded-document>
+
+</doctrine-mongo-mapping>
